gateway/requirements.txt
fastapi[standard]
uvicorn[standard] 
requests
pyjwt
graypy